requirements.txt
# Django
django==5.1.3
djangorestframework==3.15.2
django-filter==23.3
django-extensions==3.2.3
django-redis-cache==3.0.1
requests==2.32.3

 # celery
celery==5.3.4
django-celery-results==2.5.1
django-celery-beat==2.5.0

# Database
mysql-connector-python==9.1.0

# coverage & tests
coverage==7.3.2
model-bakery==1.15.0

# pre-commit, linter, formatter
pre_commit==3.4.0
flake8==6.1.0
flake8-bugbear==23.9.16
isort==5.12.0
autoflake==2.2.1
black==23.9.1
